+++ metrics/tarantool/slab.lua
@@ -12,10 +12,12 @@ local function update_slab_metrics()
     for k, v in pairs(slab_info) do
         local metric_name = 'slab_' .. k
         if not k:match('_ratio$') then
-            collectors_list[metric_name] = utils.set_gauge(metric_name, 'Slab ' .. k .. ' info', v)
+            collectors_list[metric_name] = utils.set_gauge(metric_name, 'Slab ' .. k .. ' info', v,
+                nil, nil, {default = true})
         else
             collectors_list[metric_name] =
-                utils.set_gauge(metric_name, 'Slab ' .. k .. ' info', tonumber(v:match('^([0-9%.]+)%%?$')))
+                utils.set_gauge(metric_name, 'Slab ' .. k .. ' info', tonumber(v:match('^([0-9%.]+)%%?$')),
+                nil, nil, {default = true})
         end
     end
 end