+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Handle to clear psutils metrics
 - `invoke_callbacks` option for `metrics.collect()`
 - Ability to set metainfo for collectors
+- Set `metainfo.default` to `true` for all collectors
+  from `enable_default_metrics()` and psutils collectors
 
 ### Fixed